Use a scope per request, instead of one scope per preprocessor.
If two web requests came in at the same time, there would be race conditions as
they both tried to use the same scope, in different contexts. The Rhino docs
warn this will lead to races, and indeed it does. Instead we generate a scope
request.

This is slow unfortunately, as instead of compiling say coffeescript (the
compiler itself that is) once, it does it once per web request. But at least it
is correct. The next step is to compile one scope per thread, and that will be
both fast and correct.

This also simplifies the problems of scopes being everywhere. There is now one
scope per call to preprocess-X, and most code doesn't need to be aware of it
(only code using .methods need to be).