Delete the cache directory before precompilation.
I've been getting subtle bugs as a result of this. I load up assets for tests, which is
precompilation, but they actually use the assets from development, because there
are a few with the same name and the wrong one wins. So just delete all the
existing assets so this can't happen.