Basic JS engine separation for CoffeeScript.
Implemented jsengine.clj as a layer separating things that need to
execute JS from the actual engine implementation. Created stub code for
the V8 that mimics Rhino by signature, but is a no-op for now. Rewritten
cofeescript.clj asset compiler in terms of the new jsengine.clj instead
of binding directly to Rhino.