Handle inability to load v8 better
Previously, dieter could not run on Windows even when the :rhino engine was
selected, because the v8.core namespace failed to load. This change to
dieter.v8 makes it capable of handling that situation, which allows dieter to
run on Windows as well as Mac OS X and Linux. It also allows execution, using
:rhino, when the v8wrapper library is not available.

The code reflects a significant amount of experimentation to make it stable,
since ultimately, the Clojure compiler is being second-guessed and
circumvented.